Revised after first code review by @tannewt. New keycode and keyboard-layout modules. New mouse module; not yet tested.
