+++ adafruit_hid/__init__.py
@@ -26,5 +26,14 @@
 
 TODO(description)
 
-* Author(s): Scott Shawcroft
+* Author(s): Scott Shawcroft, Dan Halbert
 """
+# Import these for convenience so that the user may import the packages classes
+# directly without needing to know the file names.
+# E.g.:
+# from adafruit_hid import Keyboard
+
+from .keyboard import Keyboard
+from .keycode import Keycode
+from .mouse import Mouse
+from .us_keyboard_layout import USKeyboardLayout
