+++ api.rst
@@ -1,7 +1,21 @@
 
-.. If you created a package, create one automodule per module in the package.
-
-.. automodule:: adafruit_hid.keyboard
+.. automodule:: adafruit_hid.Keyboard
     :members:
     :member-order: bysource
-    :synopsis: HID keyboard simulation
+    :synopsis: HID keyboard simulation.
+
+
+.. automodule:: adafruit_hid.Keycode
+    :members:
+    :member-order: bysource
+    :synopsis: HID keycode constants.
+
+.. automodule:: adafruit_hid.USKeyboardLayout
+    :members: USKeyboardLayout
+    :member-order: bysource
+    :synopsis: Write strings to a Keyboard, assuming a US keyboard layout.
+
+.. automodule:: adafruit_hid.Mouse
+    :members: Mouse
+    :member-order: bysource
+    :synopsis: HID mouse simulation.
